checking in gradlew (#37)
The Wrapper is something you should check into version control. By distributing the Wrapper with your project, anyone can work with it without needing to install Gradle beforehand. Even better, users of the build are guaranteed to use the version of Gradle that the build was designed to work with.

https://docs.gradle.org/current/userguide/gradle_wrapper.html